requirements.txt
Flask==2.0.1
Werkzeug==2.0.2
requests==2.28.2
PyJWT==2.6.0
python-dotenv==1.0.0